theme for the survey preview
