+++ projects/safe/src/lib/components/form-builder/form-builder.component.ts
@@ -16,32 +16,7 @@ import { SafeSnackBarService } from '../../services/snackbar/snackbar.service';
 import { SafeReferenceDataService } from '../../services/reference-data/reference-data.service';
 import { Form } from '../../models/form.model';
 import { renderGlobalProperties } from '../../survey/render-global-properties';
-
-/**
- * Returns a lighter version of a hexadecimal color by a specified percentage.
- *
- * @param color - The hexadecimal color to be modified.
- * @param percent - The percentage by which to lighten the color.
- * @returns A hexadecimal color that is lighter than the original color.
- */
-const lightenHexColor = (color: string, percent: number) => {
-  // Convert the color to RGB
-  const r = parseInt(color.substring(1, 3), 16);
-  const g = parseInt(color.substring(3, 5), 16);
-  const b = parseInt(color.substring(5, 7), 16);
-
-  // Lighten the color by the specified percent
-  const rl = Math.round(r + (255 - r) * (percent / 100));
-  const gl = Math.round(g + (255 - g) * (percent / 100));
-  const bl = Math.round(b + (255 - b) * (percent / 100));
-
-  // Convert the lightened color back to hex
-  const rlHex = rl.toString(16).padStart(2, '0');
-  const glHex = gl.toString(16).padStart(2, '0');
-  const blHex = bl.toString(16).padStart(2, '0');
-
-  return `#${rlHex}${glHex}${blHex}`;
-};
+import { lightenHexColor } from '../../utils/lightenHexColor';
 
 /**
  * Array containing the different types of questions.
@@ -317,11 +292,18 @@ export class SafeFormBuilderComponent implements OnInit, OnChanges {
     const primary20 = lightenHexColor(this.environment.theme.primary, 80);
 
     const defaultThemeColorsSurvey = Survey.StylesManager.ThemeColors.default;
+
+    // theme for the survey creator
     defaultThemeColorsSurvey.$primary = this.environment.theme.primary;
     defaultThemeColorsSurvey.$secondary = this.environment.theme.primary;
     defaultThemeColorsSurvey['$primary-light'] = primary10;
     defaultThemeColorsSurvey['$secondary-light'] = primary20;
     defaultThemeColorsSurvey['$secondary-back-light'] = primary10;
+
+    // theme for the survey preview
+    defaultThemeColorsSurvey['$main-color'] = this.environment.theme.primary;
+    defaultThemeColorsSurvey['$main-hover-color'] =
+      this.environment.theme.primary;
 
     Survey.StylesManager.applyTheme();
     SurveyCreator.StylesManager.applyTheme();
